refactor: created common component for messages
